Step 4 Code Cleanup
Deleted IRSpectrum.py and CompareQueryToDB.py

Renamed FormatePDFData.py to Query.py

Query.py now works correctly, though still needs to be optimized.

TODO: main.js needs to be modified to only call Query.py once.

Program is 100% functional though it is a bit slow(it's running twice) till main.js only calls Query.py once.
