+++ output/Ranked Differences.txt
@@ -1,105 +1,105 @@
-#1: 100-46-9_a.pdf
+#1: 62-53-3_b.pdf
 difference = 0.0
 
-#2: 100-46-9_c.pdf
-difference = 40.41196479406071
+#2: 62-53-3_c.pdf
+difference = 42.43547413311506
 
-#3: 108-88-3_c.pdf
-difference = 85.26553723175455
+#3: 110-86-1_b.pdf
+difference = 97.33538345512609
 
-#4: 120-72-9_b.pdf
-difference = 96.75882067785986
+#4: 100-02-7_a.pdf
+difference = 111.41388961506385
 
-#5: 108-88-3_b.pdf
-difference = 113.99238499835064
+#5: 89951-56-4.pdf
+difference = 134.27410024326946
 
-#6: 110-86-1_b.pdf
-difference = 117.25852126127353
+#6: 100-46-9_c.pdf
+difference = 134.87354283075857
 
-#7: 62-53-3_c.pdf
-difference = 122.0391760460251
+#7: 120-72-9_a.pdf
+difference = 136.04545052171474
 
-#8: 623-04-1.pdf
-difference = 126.27909099153639
-
-#9: 62-53-3_b.pdf
+#8: 100-46-9_a.pdf
 difference = 140.16008040215257
 
-#10: 5344-90-1.pdf
-difference = 141.76306221351072
+#9: 100-02-7_b.pdf
+difference = 140.99388875793144
 
-#11: 591-27-5.pdf
-difference = 142.45006263644078
+#10: 103-90-2.pdf
+difference = 148.62672723414565
 
-#12: 55-21-0.pdf
-difference = 150.38551345659502
+#11: 108-88-3_c.pdf
+difference = 157.815709734396
 
-#13: 100-51-6_a.pdf
-difference = 158.89938417416744
+#12: 120-72-9_b.pdf
+difference = 158.73609793137484
 
-#14: 620-24-6.pdf
-difference = 160.60945874825308
+#13: 108-88-3_b.pdf
+difference = 159.6159822183768
 
-#15: 623-05-2.pdf
-difference = 164.45518649280817
+#14: 98-95-3_a.pdf
+difference = 162.93555001241856
 
-#16: 100-02-7_b.pdf
-difference = 168.53378763869114
+#15: 98-95-3_c.pdf
+difference = 165.7769695445099
 
-#17: 110-85-0_c.pdf
-difference = 170.41900935869992
+#16: 620-24-6.pdf
+difference = 167.12930507009787
 
-#18: 100-02-7_a.pdf
-difference = 171.10228189860038
+#17: 50-78-2_a.pdf
+difference = 176.63473462063675
 
-#19: 90-01-7.pdf
-difference = 173.25144243163032
+#18: 623-04-1.pdf
+difference = 179.56390239684964
 
-#20: 120-72-9_a.pdf
-difference = 188.2611964422431
+#19: 55-21-0.pdf
+difference = 180.8328597559506
 
-#21: 103-90-2.pdf
-difference = 188.8680586870003
+#20: 99-05-8.pdf
+difference = 181.3395138100227
 
-#22: 100-51-6_c.pdf
-difference = 190.38099062540482
+#21: 50-78-2_b.pdf
+difference = 182.38660479230356
 
-#23: 89951-56-4.pdf
-difference = 190.7176744324409
+#22: 591-27-5.pdf
+difference = 191.37150423463083
 
-#24: 99-05-8.pdf
-difference = 206.2830246834619
+#23: 90-01-7.pdf
+difference = 191.3865000208275
 
-#25: 110-85-0_a.pdf
-difference = 213.59616714851998
+#24: 5344-90-1.pdf
+difference = 197.0523004012656
 
-#26: 50-78-2_a.pdf
-difference = 215.10861627786298
+#25: 623-05-2.pdf
+difference = 206.0020114976954
+
+#26: 110-85-0_c.pdf
+difference = 223.49190110843153
 
 #27: 110-85-0_b.pdf
-difference = 215.2480103784397
+difference = 224.28889151741836
 
-#28: 50-78-2_b.pdf
-difference = 223.26952139869474
+#28: 556-18-3.pdf
+difference = 240.62946403798736
 
-#29: 556-18-3.pdf
-difference = 243.29639709671318
+#29: 100-51-6_a.pdf
+difference = 249.9516321357425
 
-#30: 98-95-3_a.pdf
-difference = 249.31829216395084
+#30: 110-85-0_a.pdf
+difference = 260.1173559812914
 
-#31: 98-95-3_c.pdf
-difference = 254.08567474157957
+#31: 100-51-6_c.pdf
+difference = 268.1506394469736
 
 #32: 16082-33-0.pdf
-difference = 302.74509974352407
+difference = 339.7548828723435
 
 #33: 31230-17-8.pdf
-difference = 309.1760883543118
+difference = 347.2008919806547
 
 #34: 110-89-4_c.pdf
-difference = 435.52367261565587
+difference = 481.2308272930104
 
 #35: 110-89-4_b.pdf
-difference = 469.0596609155599
+difference = 515.2688664172538
